test: re-modularize contract tests
