+++ extension/src/tests/utils.rs
@@ -5,79 +5,8 @@ use crate::{
 	ContractWeights, DefaultConverter, Dispatchable, Environment as _, GetDispatchInfo, RawOrigin,
 	Weight,
 };
-use codec::Encode;
-use core::fmt::Debug;
 use frame_support::assert_ok;
-use pallet_contracts::{Code, CollectEvents, ContractExecResult, Determinism, WeightInfo};
-use std::path::Path;
-
-/// Initializing a new contract file if it does not exist.
-pub(crate) fn initialize_contract(contract_path: &str) -> Vec<u8> {
-	if !Path::new(contract_path).exists() {
-		use contract_build::*;
-		let manifest_path = ManifestPath::new("contract/Cargo.toml").unwrap();
-		let args = ExecuteArgs {
-			build_artifact: BuildArtifacts::CodeOnly,
-			build_mode: BuildMode::Debug,
-			manifest_path,
-			output_type: OutputType::Json,
-			verbosity: Verbosity::Quiet,
-			skip_wasm_validation: true,
-			..Default::default()
-		};
-		execute(args).unwrap();
-	}
-	std::fs::read(contract_path).unwrap()
-}
-
-/// Instantiating the contract.
-pub(crate) fn instantiate(func_name: &str, contract: Vec<u8>) -> AccountId {
-	let result = Contracts::bare_instantiate(
-		ALICE,
-		0,
-		GAS_LIMIT,
-		None,
-		Code::Upload(contract),
-		function_selector(func_name),
-		Default::default(),
-		DEBUG_OUTPUT,
-		CollectEvents::UnsafeCollect,
-	);
-	log::debug!("instantiate result: {result:?}");
-	let result = result.result.unwrap();
-	assert!(!result.result.did_revert());
-	result.account_id
-}
-
-/// Perform a call to a specified contract.
-pub(crate) fn call(
-	func_name: &str,
-	contract: AccountId,
-	func_id: u32,
-	input: impl Encode + Debug,
-	gas_limit: Weight,
-) -> ContractExecResult<Balance, EventRecord> {
-	log::debug!("call: func_id={func_id}, input={input:?}");
-	let result = Contracts::bare_call(
-		ALICE,
-		contract,
-		0,
-		gas_limit,
-		None,
-		[function_selector(func_name), (func_id, input.encode()).encode()].concat(),
-		DEBUG_OUTPUT,
-		CollectEvents::UnsafeCollect,
-		Determinism::Enforced,
-	);
-	log::debug!("gas consumed: {:?}", result.gas_consumed);
-	log::debug!("call result: {result:?}");
-	result
-}
-
-/// Construct the hashed bytes as a selector of function.
-pub(crate) fn function_selector(name: &str) -> Vec<u8> {
-	sp_io::hashing::blake2_256(name.as_bytes())[0..4].to_vec()
-}
+use pallet_contracts::WeightInfo;
 
 // Weight charged for calling into the runtime from a contract.
 pub(crate) fn overhead_weight(input_len: u32) -> Weight {
